test(comment): command repository 테스트 추가 및 분리 - v1에 대한 유닛테스트, 동시성 테스트 추가
