+++ src/test/java/com/simpleboard/board/board/domain/comment/repository/CommentCommandRepositoryConcurrencyTest.java
@@ -3,14 +3,8 @@ package com.simpleboard.board.board.domain.comment.repository;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.simpleboard.board.board.domain.comment.dto.CommentCreateParams;
-import com.simpleboard.board.board.domain.comment.dto.CommentDeleteParams;
 import com.simpleboard.board.board.domain.comment.entity.Comment;
-import com.simpleboard.board.board.domain.comment.entity.GuestComment;
-import com.simpleboard.board.board.domain.comment.entity.MemberComment;
-import com.simpleboard.board.board.domain.comment.vo.CommentState;
 import com.simpleboard.board.board.domain.comment.vo.CommentType;
-import com.simpleboard.board.board.domain.common.vo.Visitor;
-import com.simpleboard.board.board.domain.common.vo.VisitorType;
 import com.simpleboard.board.board.infrastructure.jpa.converter.CommentConverter;
 import com.simpleboard.board.board.infrastructure.jpa.entity.CommentEntity;
 import com.simpleboard.board.board.infrastructure.jpa.repository.CommentCommandRepositoryImpl;
@@ -34,109 +28,14 @@ import org.springframework.test.context.ActiveProfiles;
  * <p>H2 DB를 활용한 JPA 레포지토리 테스트
  */
 @DataJpaTest
-@ActiveProfiles("test")
+@ActiveProfiles("test-unit")
 @Import({CommentCommandRepositoryImpl.class, CommentConverter.class, JpaAuditTestConfig.class})
-class CommentCommandRepositoryTest {
+class CommentCommandRepositoryConcurrencyTest {
 
-  private static final Logger log = LoggerFactory.getLogger(CommentCommandRepositoryTest.class);
+  private static final Logger log =
+      LoggerFactory.getLogger(CommentCommandRepositoryConcurrencyTest.class);
   @Autowired CommentCommandRepository repository;
   @Autowired CommentEntityRepository entityRepository;
-
-  @Nested
-  @DisplayName("save")
-  class SaveTests {
-
-    @Test
-    @DisplayName("GuestComment 를 저장하면 ID가 채워지고 필드가 보존된다")
-    void save_guest() {
-      // given
-      Comment comment = Comment.write(guestParams());
-
-      // when
-      Comment saved = repository.save(comment);
-
-      // then
-      assertThat(saved).isInstanceOf(GuestComment.class);
-      assertThat(saved.getId()).isNotNull();
-      assertThat(saved.getPostId()).isEqualTo(100L);
-      assertThat(saved.getParentId()).isNull();
-      assertThat(saved.getContent()).isEqualTo("hello-guest");
-      assertThat(saved.getCommentState()).isEqualTo(CommentState.ACTIVATE);
-
-      GuestComment gc = (GuestComment) saved;
-      assertThat(gc.getNickname()).isEqualTo("게스트닉");
-      assertThat(gc.getPassword()).isEqualTo("pass-1234");
-    }
-
-    @Test
-    @DisplayName("MemberComment 를 저장하면 ID가 채워지고 필드가 보존된다")
-    void save_member() {
-      // given
-      Comment comment = Comment.write(memberParams());
-
-      // when
-      Comment saved = repository.save(comment);
-
-      // then
-      assertThat(saved).isInstanceOf(MemberComment.class);
-      assertThat(saved.getId()).isNotNull();
-      assertThat(saved.getPostId()).isEqualTo(200L);
-      assertThat(saved.getParentId()).isEqualTo(10L);
-      assertThat(saved.getContent()).isEqualTo("hello-member");
-      assertThat(saved.getCommentState()).isEqualTo(CommentState.ACTIVATE);
-
-      MemberComment mc = (MemberComment) saved;
-      assertThat(mc.getWriterId()).isEqualTo(42L);
-    }
-  }
-
-  @Nested
-  @DisplayName("findById")
-  class FindTests {
-
-    @Test
-    @DisplayName("저장 후 ID로 조회하면 동일 타입/값으로 변환되어 돌아온다 (Guest)")
-    void find_guest_roundtrip() {
-      // given
-      Comment saved = repository.save(Comment.write(guestParams()));
-
-      // when
-      Optional<Comment> foundOpt = repository.findById(saved.getId());
-
-      // then
-      assertThat(foundOpt).isPresent();
-      Comment found = foundOpt.get();
-      assertThat(found).isInstanceOf(GuestComment.class);
-      assertThat(found.getId()).isEqualTo(saved.getId());
-      assertThat(found.getContent()).isEqualTo("hello-guest");
-      assertThat(((GuestComment) found).getNickname()).isEqualTo("게스트닉");
-    }
-
-    @Test
-    @DisplayName("저장 후 삭제 API 시나리오를 흉내내면 상태가 DELETED로 바뀐다 (Member)")
-    void delete_member_flow_like_service() {
-      // given: 저장
-      MemberComment saved = (MemberComment) repository.save(Comment.write(memberParams()));
-
-      // when: 도메인 삭제 로직 실행 (권한자)
-      Visitor writer =
-          Visitor.builder().type(VisitorType.MEMBER).memberId(42L).vId("vid-xxx").build();
-      saved.deleteComment(writer, CommentDeleteParams.builder().password(null).build());
-      Comment after = repository.save(saved);
-
-      // then
-      assertThat(after.getCommentState()).isEqualTo(CommentState.DELETED);
-      // 조회해서도 삭제 상태 유지 확인
-      Comment found = repository.findById(after.getId()).orElseThrow();
-      assertThat(found.getCommentState()).isEqualTo(CommentState.DELETED);
-    }
-
-    @Test
-    @DisplayName("존재하지 않는 ID 조회시 Optional.empty")
-    void not_found() {
-      assertThat(repository.findById(-999L)).isEmpty();
-    }
-  }
 
   @Nested
   @DisplayName("Insert Concurrency")
@@ -264,16 +163,6 @@ class CommentCommandRepositoryTest {
         .commentType(CommentType.GUEST)
         .nickname("child")
         .password("pass-1234")
-        .build();
-  }
-
-  private CommentCreateParams memberParams() {
-    return CommentCreateParams.builder()
-        .postId(200L)
-        .parentCommentId(10L)
-        .content("hello-member")
-        .commentType(CommentType.MEMBER)
-        .writerId(42L)
         .build();
   }
 }
